+++ foundry/settings.py
@@ -31,7 +31,7 @@ TEMPLATE_DEBUG = DEBUG
 
 DATABASES = {
     'default': {
-        'ENGINE': 'django.db.backends.postgresql_psycopg2', # Add 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
+        'ENGINE': 'django.contrib.gis.db.backends.postgis', # Add 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
         'NAME': 'foundry', # Or path to database file if using sqlite3.
         'USER': 'foundry', # Not used with sqlite3.
         'PASSWORD': 'foundry', # Not used with sqlite3.
@@ -83,7 +83,7 @@ MIDDLEWARE_CLASSES = (
     'foundry.middleware.AgeGateway',
     'django.contrib.messages.middleware.MessageMiddleware',
     'likes.middleware.SecretBallotUserIpUseragentMiddleware',
-    'foundry.middleware.PaginationMiddleware',
+    'pagination.middleware.PaginationMiddleware',
     'foundry.middleware.VerboseRequestMeta',                    
     'django.middleware.transaction.TransactionMiddleware',
 )